Remove safe-mode
